refactor: delete streaming lifecycle forwarders
